update gitea solution
