--- ThreeBotPackages/tfgrid_solutions/gitea/wiki/gitlab.md
@@ -1,21 +1,0 @@
-## GitLab
-A web-based DevOps lifecycle tool that provides a Git-repository manager providing wiki, issue-tracking
-and CI/CD pipeline features, using an open-source license, developed by GitLab Inc.
-
-The software was created by Ukrainians Dmitriy Zaporozhets and Valery Sizov, and is used by
-several large tech companies including IBM, Sony, Jülich Research Center, NASA, Alibaba, Oracle,
-Invincea, O’Reilly Media, Leibniz-Rechenzentrum (LRZ), CERN, European XFEL, GNOME Foundation,
-Boeing, Autodata, and SpaceX.
-
-The code was originally written in Ruby, with some parts later rewritten in Go, initially as a
-source code management solution to collaborate with his team on software development.
-
-It later evolved to an integrated solution covering the software development life cycle,
-and then to the whole DevOps life cycle. The current technology stack includes Go,
-Ruby on Rails and Vue.js.
-
-It follows an open-core development model where the core functionality is released under
-an open source (MIT) license while the additional functionality is under a proprietary license.
-
-GitLab is considered as the first Ukrainian unicorn valued more than $1 billion.
-
